Rename LogoutButtonRename to LogoutButton component, test and story
