+++ src/components/LogoutButton.spec.js
